refacto: split types by categories
