+++ src/types/http.ts
@@ -1,0 +1,6 @@
+export type BearerToken = `Bearer ${string}`;
+
+export interface RequestHeaders {
+  Authorization: BearerToken;
+  "Content-Type": "application/json";
+}
